Add an abstract concept of a `PaneItem`
Until now, there was not a whole lot of documentation of what methods are considered “special” on pane items. There is a loose interface, and most of these methods are considered optional, but it's not accurate to type `PaneItem` simply as `object`, so this is a step in the right direction.