Change return for `CommandRegistryListener` from `void` to `unknown`…
…to make it easier to use.

The actual `CommandRegistry` doesn't care whether you return stuff from your handler or not. (That would've been painful back in the CoffeeScript days!)

The intent of `void` is that there is no _expected_ return value; but it also has the effect of making TypeScript complain when your command listeners _happen to_ return a value.

Maybe you're using the same function in several different contexts! Maybe we should be more forgiving.