Add the `LanguageMode` interface and the associated `TextBuffer` methods
This might open a very small Pandora's Box. It's arguably useful to document at least _some_ of the `LanguageMode` API in order to reduce the amount of `@ts-ignore`-ing that someone must do.

That said, it's most definitely not worth our time to try to document all possible methods that can be called on a `LanguageMode`. For instance, if a package needs to call a method specific to a `WASMTreeSitterLanguageMode`, that's fine… but they'll have to sprinkle `@ts-ignore`s around.